+++ packages/sui-analytics/src/index.js
@@ -6,6 +6,13 @@ export default {
     window.analytics.reset(...args)
   },
   track: (...args) => {
-    window.analytics.track(...args)
+    window.analytics.ready(function() {
+      const [event, properties, options = {}, fn] = args
+      options.traits = {
+        anonymousId: window.analytics.user().anonymousId(),
+        userId: window.analytics.user().id()
+      }
+      window.analytics.track(event, properties, options, fn)
+    })
   }
 }