+++ packages/sui-bundler/shared/config.js
@@ -1,0 +1,5 @@
+const {config} = require('./')
+
+const {sourcemaps} = config
+
+exports.sourceMap = sourcemaps && sourcemaps.prod ? sourcemaps.prod : 'none'