+++ packages/sui-bundler/shared/uglify.js
@@ -1,7 +1,0 @@
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
-
-module.exports = new UglifyJsPlugin({
-  cache: true,
-  parallel: true,
-  sourceMap: true
-})