+++ packages/sui-bundler/webpack.config.lib.js
@@ -1,14 +1,16 @@
 const webpack = require('webpack')
 const {cleanList, envVars, MAIN_ENTRY_POINT, config} = require('./shared')
-const uglifyJsPlugin = require('./shared/uglify')
+const minifyJs = require('./shared/minify-js')
 const definePlugin = require('./shared/define')
 const babelRules = require('./shared/module-rules-babel')
+const {sourceMap} = require('./shared/config')
+const parseAlias = require('./shared/parse-alias')
 require('./shared/shims')
 
 module.exports = {
   mode: 'production',
   resolve: {
-    alias: config.alias,
+    alias: parseAlias(config.alias),
     extensions: ['*', '.js', '.jsx', '.json']
   },
   entry: config.vendor
@@ -24,7 +26,7 @@ module.exports = {
     filename: 'index.js'
   },
   optimization: {
-    minimizer: [uglifyJsPlugin]
+    minimizer: [minifyJs(sourceMap)]
   },
   plugins: cleanList([
     new webpack.HashedModuleIdsPlugin(),