+++ packages/sui-js/src/ua-parser/index.js
@@ -2,8 +2,11 @@ import bowser from 'bowser'
 
 export const stats = userAgent => {
   const ua = bowser._detect(userAgent)
+
   return {
     isMobile: !!ua.mobile,
-    osName: ua.osname
+    osName: ua.osname,
+    browserName: ua.name,
+    browserVersion: ua.version
   }
 }