+++ packages/sui-js/test/stringSpec.js
@@ -1,6 +1,10 @@
 /* eslint-env mocha */
 import {expect} from 'chai'
-import {parseQueryString, toQueryString} from '../src/string/index'
+import {
+  parseQueryString,
+  toQueryString,
+  fromArrayToCommaQueryString
+} from '../src/string/index'
 
 describe('@s-ui/js', () => {
   describe('string:toQueryString', () => {
@@ -15,6 +19,13 @@ describe('@s-ui/js', () => {
       const query = '?a=1&b=test'
       const params = {a: '1', b: 'test'}
       expect(parseQueryString(query)).to.deep.equal(params)
+    })
+  })
+  describe('string:fromArrayToCommaQueryString', () => {
+    it('should convert params array to comma separated object params', () => {
+      const paramsArray = {a: 1, b: 2, c: [3, 4]}
+      const params = 'a=1&b=2&c=3,4'
+      expect(fromArrayToCommaQueryString(paramsArray)).to.deep.equal(params)
     })
   })
 })